프로젝트 투두리스트 폼 생성 -> 기존 작업 변환 작업 중

+++ src/main/java/com/hta2405/unite/dto/ProjectDetailDTO.java
@@ -1,6 +1,9 @@
 package com.hta2405.unite.dto;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.ToString;
 
 @Getter
 //@Setter
@@ -13,8 +16,9 @@ public class ProjectDetailDTO {
     private String managerId;
     private String participantNames; // 참여자 목록
     private String memberDesign; //진행률 - 업무
-    private int memberProgressRate;  //진행률 - 진행률
+    private double memberProgressRate;  //진행률 - 진행률
     private boolean isManager = false;
+
     public void setIsManager(boolean isManager) {
         this.isManager = isManager;
     }
